+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -134,6 +134,7 @@ The version number for this package has increased due to a version update of a r
 - Fix yamato error : check vfx manager on domain reload instead of vfx import.
 - Filter out unrelevant events from event desc while compiling
 - Missing Packing.hlsl include while using an unlit shadergraph.
+- Handle correctly disabled alphaTreshold material slot in shaderGraph.
 
 ## [7.1.1] - 2019-09-05
 ### Added